Move AdditionalValidation to common package for reusability in V1 flow
